Rename core to tokens
