+++ Backpack/src/main/java/net/skyscanner/backpack/gradient/BpkGradients.kt
@@ -1,8 +1,9 @@
-package net.skyscanner.backpack
+package net.skyscanner.backpack.gradient
 
 import android.content.Context
 import android.graphics.drawable.GradientDrawable
 import android.support.v4.content.ContextCompat
+import net.skyscanner.backpack.R
 
 /**
  * Backpack gradient utility
