+++ app/src/androidTest/java/net/skyscanner/backpack/gradient/BpkGradientsTest.kt
@@ -1,10 +1,11 @@
-package net.skyscanner.backpack
+package net.skyscanner.backpack.gradient
 
 import android.content.Context
 import android.graphics.drawable.GradientDrawable
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
 import android.view.View
+import net.skyscanner.backpack.BpkSnapshotTest
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
