--- bpk-avd.ini
@@ -1,22 +1,0 @@
-PlayStore.enabled=false
-abi.type=armeabi-v7a
-avd.ini.encoding=UTF-8
-hw.lcd.height=1920
-hw.lcd.width=1080
-hw.sdCard=yes
-hw.sensors.orientation=yes
-hw.sensors.proximity=yes
-hw.trackBall=no
-image.sysdir.1=system-images/android-27/google_apis/x86/
-tag.display=Google APIs
-tag.id=google_apis
-vm.heapSize=512
-hw.ramSize=1536
-disk.dataPartition.size=800M
-hw.gpu.enabled=yes
-hw.gpu.mode=auto
-
-
-hw.lcd.density=240
-hw.ramSize=512
-vm.heapSize=48
